requirements.txt
boto==2.48.0
babel==2.5.1
celery==4.1.0
channels==1.1.8
dj-database-url==0.4.2
django-constance==2.0.0
django-filter==1.1.0
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-picklefield==1.0.0
django-storages==1.6.5
Django==1.11.6
djangorestframework==3.7.1
djoser==1.0.1
git+git://github.com/Brogency/djangorestframework-camel-case@a6a448360941130678b62aebed5d5d09ba1214bb#egg=djangorestframework-camel-case
gunicorn==19.7.1
Pillow==4.3.0
psycopg2==2.7.3.2
python-postmark==0.4.12
pytz==2017.2
raven==6.2.1
redis==2.10.6
requests==2.18.4
whitenoise==3.3.1

django-extensions==1.9.6
ipdb
ipdbplugin
ipython==6.2.1

django-db-email-backend==0.3
coverage==4.4.1
django-nose==1.4.5
factory_boy==2.9.2
Faker==0.8.6
rednose==1.2.2

